scale Nthermal and stride by the size of the system

+++ sim.hpp
@@ -12,7 +12,7 @@
 class Sim
 {
 public:
-    Sim(size_t N, size_t Nmeasure, size_t Nthermal, size_t stride, std::string const& filename, double temperature, double kb, double J, Vec H);
+    Sim(size_t N, size_t Nmeasure, double Nthermal, double stride, std::string const& filename, double temperature, double kb, double J, Vec H);
     Sim(Config const& params);
 
     void run();
